+++ pyapireference/ui/button_with_extra_options.py
@@ -6,10 +6,10 @@ from PyQt5.QtGui import QCursor, QFont, QFontMetrics
 from typing import List, Tuple, Callable
 
 if __name__ == "__main__":
-    raise RuntimeError("button_with_extra_options.py requires get_text_size from extra.py which is outside this folder, you can't run this script as main")
+    raise RuntimeError("button_with_extra_options.py requires get_text_size from pyapireference.extra.py which is outside this folder, you can't run this script as main")
 else:
-    import GUI.collapsible_widget_resources
-    from extra import get_text_size, create_menu
+    import pyapireference.ui.collapsible_widget_resources
+    from pyapireference.extra import get_text_size, create_menu
 
 class MainWindow(QMainWindow):
 	def __init__(self, *args, **kwargs):
@@ -49,7 +49,7 @@ class ButtonWithExtraOptions(QWidget):
 
 		for action_name, action_callback in self.actions:
 			if isinstance(action_callback, dict):
-				menu.addMenu(create_menu(action_name, action_callback, parent=self))
+				menu.addMenu(create_menu(action_callback, action_name, parent=self))
 				continue
 
 			menu.addAction(action_name, action_callback)