+++ pyapireference/ui/markdown_text_edit.py
@@ -1,5 +1,6 @@
 from PyQt5.QtWidgets import QApplication, QTextEdit
 from PyQt5.QtGui import QTextOption
+from PyQt5.QtCore import QEvent
 
 class MarkdownTextEdit(QTextEdit):
 	def __init__(self, *args, **kwargs):